+++ packages/common-ethereum/CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
 ## [3.8.0] - 2024-06-07
 ### Added
 - Improve ABI validation