+++ packages/common-ethereum/CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.8.3] - 2024-06-11
 ### Fixed
 - Fix `generateAbis` when datasource without assets
 
@@ -166,7 +167,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Sync with main sdk (#14)
 
 ## [0.1.0] - 2022-10-31
-[Unreleased]: https://github.com/subquery/subql-ethereum/compare/common-ethereum/3.8.2...HEAD
+[Unreleased]: https://github.com/subquery/subql-ethereum/compare/common-ethereum/3.8.3...HEAD
+[3.8.3]: https://github.com/subquery/subql-ethereum/compare/common-ethereum/3.8.2...common-ethereum/3.8.3
 [3.8.2]: https://github.com/subquery/subql-ethereum/compare/common-ethereum/3.8.1...common-ethereum/3.8.2
 [3.8.1]: https://github.com/subquery/subql-ethereum/compare/common-ethereum/3.7.0...common-ethereum/3.8.1
 [3.7.0]: https://github.com/subquery/subql-ethereum/compare/common-ethereum/3.6.1...common-ethereum/3.7.0